+++ vembrane/modules/common.py
@@ -1,0 +1,18 @@
+import ast
+from vembrane.errors import InvalidExpression
+
+
+def check_filter_expression(expression: str,) -> str:
+    if ".__" in expression:
+        raise InvalidExpression(expression, "The expression must not contain '.__'")
+    try:
+        tree = ast.parse(expression, mode="eval")
+        if isinstance(tree.body, (ast.BoolOp, ast.Compare)):
+            return expression
+        else:
+            # TODO possibly check for ast.Call, func return type
+            return expression
+    except SyntaxError:
+        raise InvalidExpression(
+            expression, "The expression has to be syntactically correct."
+        )
